requirements.txt
Django==4.1.1
djangorestframework==3.13.1
psycopg2-binary==2.9.1
APScheduler==3.9.1